Created a class for person/people function
